+++ src/test/java/tests/DynamicTestData.java
@@ -1,15 +1,13 @@
-package utils;
+package tests;
 
 import com.github.javafaker.Faker;
 
-import java.security.SecureRandom;
 import java.text.SimpleDateFormat;
-import java.util.Arrays;
 import java.util.Date;
 import java.util.Locale;
 import java.util.concurrent.ThreadLocalRandom;
 
-public class DynamicDataUtils {
+public class DynamicTestData {
     public static void main(String[] args) {
         String[] genders = {"Male", "Female", "Other"};
         System.out.println(getRandomItemFromArray(genders));
